+++ src/lib/eiotas_link.c
@@ -21,7 +21,7 @@
 
 void eiotas_link_free(Eiotas_Link *link)
 {
-    DBG("Link free 0x%X",link);
+    DBG("Link free 0x%X",(unsigned int)link);
     // TODO
 }
